+++ src/lib/file-parser.ts
@@ -8,11 +8,12 @@ import * as url from './helpers/url';
 import * as steam from './helpers/steam';
 import * as paths from "../paths";
 import * as _ from 'lodash';
-import * as glob from 'glob';
+import { globPromise } from './helpers/glob/promise';
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as os from 'os';
-import {getPath} from 'windows-shortcuts-ps';
+import * as Sentry from '@sentry/electron';
+import { getPath } from 'windows-shortcuts-ps';
 
 
 export class FileParser {
@@ -89,13 +90,14 @@ export class FileParser {
       .then(() => {
         let preParser = new VariableParser({ left: '${', right: '}' });
         for (let i = 0; i < configs.length; i++) {
-          let isGlobParser:boolean = configs[i].parserType!=='Steam';
-
+          let isSteamParser:boolean = configs[i].parserType=='Steam';
+          let isImporterParser:boolean = ['Epic'].includes(configs[i].parserType);
+          let isGlobParser:boolean = ['Glob','Glob-regex'].includes(configs[i].parserType);
           // Parse environment variables on rom directory, start in path, executable path
           configs[i].steamDirectory = preParser.setInput(configs[i].steamDirectory).parse() ? preParser.replaceVariables((variable) => {
             return this.getEnvironmentVariable(variable as EnvironmentVariables,settings).trim()
           }) : null;
-          if(isGlobParser){
+          if(isGlobParser) {
             configs[i].romDirectory = preParser.setInput(configs[i].romDirectory).parse() ? preParser.replaceVariables((variable) => {
               return this.getEnvironmentVariable(variable as EnvironmentVariables,settings).trim()
             }) : null;
@@ -128,7 +130,9 @@ export class FileParser {
 
         let promises: Promise<ParsedData>[] = [];
         for(let i=0; i<configs.length;i++) {
-          let isGlobParser:boolean = configs[i].parserType!=='Steam';
+          let isSteamParser:boolean = configs[i].parserType=='Steam';
+          let isImporterParser:boolean = ['Epic'].includes(configs[i].parserType);
+          let isGlobParser:boolean = ['Glob','Glob-regex'].includes(configs[i].parserType);
           let parser = this.getParserInfo(configs[i].parserType);
           if (parser) {
             if (parser.inputs !== undefined) {
@@ -143,7 +147,7 @@ export class FileParser {
             let userFilter = preParser.setInput(configs[i].userAccounts.specifiedAccounts).parse() ? _.uniq(preParser.extractVariables(data => null)) : [];
             filteredAccounts.push(this.filterUserAccounts(steamDirectories[i].data, userFilter, configs[i].steamDirectory, configs[i].userAccounts.skipWithMissingDataDir));
             totalUserAccountsFound+=filteredAccounts[filteredAccounts.length-1].found.length;
-            let directories = isGlobParser? [configs[i].romDirectory] : filteredAccounts[i].found.map((account: userAccountData)=>path.join(configs[i].steamDirectory,'userdata',account.accountID));
+            let directories = isGlobParser ? [configs[i].romDirectory] : filteredAccounts[i].found.map((account: userAccountData)=>path.join(configs[i].steamDirectory,'userdata',account.accountID));
             promises.push(this.availableParsers[configs[i].parserType].execute(directories, configs[i].parserInputs, this.globCache));
           }
           else
@@ -163,7 +167,9 @@ export class FileParser {
         let localIconPromises: Promise<void>[] = [];
         let vParser = new VariableParser({ left: '${', right: '}' });
         for (let i = 0; i < configs.length; i++) {
-          let isGlobParser:boolean = configs[i].parserType !=='Steam';
+          let isSteamParser:boolean = configs[i].parserType=='Steam';
+          let isImporterParser:boolean = ['Epic'].includes(configs[i].parserType);
+          let isGlobParser:boolean = ['Glob','Glob-regex'].includes(configs[i].parserType);
           if (isGlobParser && configs[i].titleFromVariable.tryToMatchTitle)
             this.tryToReplaceTitlesWithVariables(data[i], configs[i], vParser);
 
@@ -206,7 +212,10 @@ export class FileParser {
             if(isGlobParser) {
               executableLocation = configs[i].executable.path ? configs[i].executable.path : data[i].success[j].filePath;
               startInDir = configs[i].startInDirectory.length > 0 ? configs[i].startInDirectory : path.dirname(executableLocation);
-            } else {
+            } else if(isImporterParser){
+              executableLocation = data[i].success[j].filePath;
+              startInDir = path.dirname(executableLocation);
+            } else if(isSteamParser) {
               executableLocation = data[i].success[j].extractedAppId;
             }
 
@@ -249,7 +258,7 @@ export class FileParser {
               lastFile.finalTitle = exceptions.newTitle;
             } else {
               lastFile.finalTitle = vParser.setInput(configs[i].titleModifier).parse() ? vParser.replaceVariables((variable) => {
-                return this.getVariable(variable as AllVariables, variableData).trim();
+                  return this.getVariable(variable as AllVariables, variableData).trim();
               }) : '';
             }
             variableData.finalTitle = lastFile.finalTitle;
@@ -273,10 +282,12 @@ export class FileParser {
               return this.getVariable(variable as AllVariables, variableData);
             })) : [];
           }
+
+
+          let extRegex = /png|tga|jpg|jpeg/i;
           defaultImagePromises.push(this.resolveFieldGlobs('defaultImage', configs[i],settings, parsedConfigs[i], vParser).then((data) => {
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedDefaultImages = data.resolvedGlobs[j];
-              let extRegex = /png|tga|jpg|jpeg/i;
               for (let k = 0; k < data.resolvedFiles[j].length; k++) {
                 const item = data.resolvedFiles[j][k];
                 if (extRegex.test(path.extname(item))) {
@@ -289,7 +300,6 @@ export class FileParser {
           defaultTallImagePromises.push(this.resolveFieldGlobs('defaultTallImage',configs[i],settings,parsedConfigs[i],vParser).then((data)=>{
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedDefaultTallImages = data.resolvedGlobs[j];
-              let extRegex = /png|tga|jpg|jpeg/i;
               for (let k = 0; k < data.resolvedFiles[j].length; k++) {
                 const item = data.resolvedFiles[j][k];
                 if (extRegex.test(path.extname(item))) {
@@ -302,7 +312,6 @@ export class FileParser {
           defaultHeroImagePromises.push(this.resolveFieldGlobs('defaultHeroImage',configs[i],settings,parsedConfigs[i],vParser).then((data)=>{
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedDefaultHeroImages = data.resolvedGlobs[j];
-              let extRegex = /png|tga|jpg|jpeg/i;
               for (let k = 0; k < data.resolvedFiles[j].length; k++) {
                 const item = data.resolvedFiles[j][k];
                 if (extRegex.test(path.extname(item))) {
@@ -315,7 +324,6 @@ export class FileParser {
           defaultLogoImagePromises.push(this.resolveFieldGlobs('defaultLogoImage',configs[i],settings,parsedConfigs[i],vParser).then((data)=>{
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedDefaultLogoImages = data.resolvedGlobs[j];
-              let extRegex = /png|tga|jpg|jpeg/i;
               for (let k = 0; k < data.resolvedFiles[j].length; k++) {
                 const item = data.resolvedFiles[j][k];
                 if (extRegex.test(path.extname(item))) {
@@ -329,8 +337,6 @@ export class FileParser {
           localImagePromises.push(this.resolveFieldGlobs('localImages', configs[i],settings, parsedConfigs[i], vParser).then((data) => {
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedLocalImages = data.resolvedGlobs[j];
-
-              let extRegex = /png|tga|jpg|jpeg/i;
               data.parsedConfig.files[j].localImages = data.resolvedFiles[j].filter((item) => {
                 return extRegex.test(path.extname(item));
               }).map((item) => {
@@ -341,8 +347,6 @@ export class FileParser {
           localTallImagePromises.push(this.resolveFieldGlobs('localTallImages', configs[i],settings, parsedConfigs[i], vParser).then((data) => {
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedLocalTallImages = data.resolvedGlobs[j];
-
-              let extRegex = /png|tga|jpg|jpeg/i;
               data.parsedConfig.files[j].localTallImages = data.resolvedFiles[j].filter((item) => {
                 return extRegex.test(path.extname(item));
               }).map((item) => {
@@ -353,8 +357,6 @@ export class FileParser {
           localHeroImagePromises.push(this.resolveFieldGlobs('localHeroImages', configs[i],settings, parsedConfigs[i], vParser).then((data) => {
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedLocalHeroImages = data.resolvedGlobs[j];
-
-              let extRegex = /png|tga|jpg|jpeg/i;
               data.parsedConfig.files[j].localHeroImages = data.resolvedFiles[j].filter((item) => {
                 return extRegex.test(path.extname(item));
               }).map((item) => {
@@ -365,8 +367,6 @@ export class FileParser {
           localLogoImagePromises.push(this.resolveFieldGlobs('localLogoImages', configs[i],settings, parsedConfigs[i], vParser).then((data) => {
             for (let j = 0; j < data.parsedConfig.files.length; j++) {
               data.parsedConfig.files[j].resolvedLocalLogoImages = data.resolvedGlobs[j];
-
-              let extRegex = /png|tga|jpg|jpeg/i;
               data.parsedConfig.files[j].localLogoImages = data.resolvedFiles[j].filter((item) => {
                 return extRegex.test(path.extname(item));
               }).map((item) => {
@@ -479,7 +479,7 @@ export class FileParser {
           }) : '').replace(/\\/g, '/');
 
           resolvedGlobs[i].push(replacedGlob);
-          promises.push(this.globPromise(replacedGlob, { silent: true, dot: true, realpath: true, cwd: cwd, cache: this.globCache }).then((files) => {
+          promises.push(globPromise(replacedGlob, { silent: true, dot: true, realpath: true, cwd: cwd, cache: this.globCache }).then((files) => {
             resolvedFiles[i] = files;
           }));
         }
@@ -517,7 +517,7 @@ export class FileParser {
               }
             }
             if (secondaryMatch !== undefined) {
-              return this.globPromise(secondaryMatch, { silent: true, dot: true, realpath: true, cwd: cwd, cache: this.globCache }).then((files) => {
+              return globPromise(secondaryMatch, { silent: true, dot: true, realpath: true, cwd: cwd, cache: this.globCache }).then((files) => {
                 return resolvedFiles[i].concat(files);
               });
             }
@@ -547,6 +547,9 @@ export class FileParser {
         break;
       case 'RETROARCHPATH':
         output=settings.environmentVariables.retroarchPath;
+        break;
+      case 'RACORES':
+        output=settings.environmentVariables.raCoresDirectory;
         break;
       case 'LOCALIMAGESDIR':
         output=settings.environmentVariables.localImagesDirectory;
@@ -614,6 +617,9 @@ export class FileParser {
         break;
       case 'RETROARCHPATH':
         output=data.retroarchPath;
+        break;
+      case 'RACORES':
+        output=data.raCoresDirectory;
         break;
       case 'LOCALIMAGESDIR':
         output=data.localImagesDirectory;
@@ -700,14 +706,10 @@ export class FileParser {
               }
             }
           }
-          // TODO REPLACE WITH EXCEPTIONS
-          // if(this.argumentVariableData[output]!==undefined) {
-          //   return this.argumentVariableData[variable][data.extractedTitle] || '';
-          // }
         }
         break;
     }
-    return output;
+    return output || '';
   }
 
   private makeVariableData(config: UserConfiguration, settings: AppSettings, file: ParsedUserConfigurationFile) {
@@ -722,6 +724,7 @@ export class FileParser {
       romDirectory: config.romDirectory,
       steamDirectoryGlobal: settings.environmentVariables.steamDirectory,
       retroarchPath: settings.environmentVariables.retroarchPath,
+      raCoresDirectory: settings.environmentVariables.raCoresDirectory,
       localImagesDirectory: settings.environmentVariables.localImagesDirectory
     }
   }
@@ -733,17 +736,5 @@ export class FileParser {
     } catch (e) {
       return false;
     }
-  }
-
-  private globPromise(pattern: string, options: glob.IOptions) {
-    return new Promise<string[]>((resolve, reject) => {
-      glob(pattern, options, (err, files) => {
-        if (err)
-          reject(err);
-        else {
-          resolve(files);
-        }
-      });
-    });
   }
 }