+++ src/lib/vdf-screenshots-file.ts
@@ -4,7 +4,7 @@ import { VDF_Error } from './vdf-error';
 import { APP } from '../variables';
 import { Bluebird } from './zone-bluebird';
 import * as genericParser from '@node-steam/vdf';
-import * as glob from './helpers/glob';
+import { globPromise } from './helpers/glob';
 import * as file from './helpers/file';
 import * as ids from './helpers/steam';
 import * as _ from "lodash";
@@ -86,7 +86,7 @@ export class VDF_ScreenshotsFile {
       let screenshotsData: VDF_ScreenshotsData = this.data;
       for (let j=0; j < this.extraneous.length; j++) {
         let exAppId = this.extraneous[j]
-        promises.push(glob.promise(`${exAppId}.*`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
+        promises.push(globPromise(`${exAppId}.*`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
           let errors: Error[] = [];
           for (let i = 0; i < files.length; i++) {
             try {
@@ -104,7 +104,7 @@ export class VDF_ScreenshotsFile {
       }
       for (const appId in screenshotsData) {
         if (screenshotsData[appId] === undefined) {
-          promises.push(glob.promise(`${appId}.*`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
+          promises.push(globPromise(`${appId}.*`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
             let errors: Error[] = [];
             for (let i = 0; i < files.length; i++) {
               try {
@@ -146,7 +146,7 @@ export class VDF_ScreenshotsFile {
                 }).then((buffer) => {
                   return fs.outputFile(path.join(this.gridDirectory, `${appId}.${ids.map_ext[""+ext]||ext}`), buffer).then(() => {
                     screenshotsData[appId] = data.title;
-                    glob.promise(`${appId}.!(json)`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
+                    globPromise(`${appId}.!(json)`, { silent: true, dot: true, cwd: this.gridDirectory, absolute: true }).then((files) => {
                       let errors: Error[] = [];
                       for (let i = 0; i < files.length; i++) {
                         if(_.last(files[i].split('.'))!==(ids.map_ext[""+ext]||ext)) {