+++ src/renderer/components/parsers.component.ts
@@ -29,6 +29,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
   private userForm: FormGroup;
   private formChanges: Subscription = new Subscription();
 
+
   constructor(
     private parsersService: ParsersService,
     private loggerService: LoggerService,
@@ -78,7 +79,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
           }
         }),
         executable: new NestedFormElement.Group({
-          isHidden: () => this.isHiddenIfSteamParser(),
+          isHidden: () => this.isHiddenIfNotRomsParser(),
           label: this.lang.label.executableLocation,
           children: {
             path: new NestedFormElement.Path({
@@ -93,7 +94,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
               text: this.lang.text.shortcut_passthrough
             }),
             appendArgsToExecutable: new NestedFormElement.Toggle({
-              isHidden: () => this.isHiddenIfEither(),
+              isHidden: () => this.isHiddenIfAny(),
               text: this.lang.text.appendArgsToExecutable
             })
           },
@@ -113,7 +114,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
           }
         }),
         executableModifier: new NestedFormElement.Input({
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           highlight: this.highlight.bind(this),
           label: this.lang.label.executableModifier,
           onValidate: (self, path) => this.parsersService.validate(path[0] as keyof UserConfiguration, self.value),
@@ -123,7 +124,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
           }
         }),
         romDirectory: new NestedFormElement.Path({
-          isHidden: () => this.isHiddenIfSteamParser(),
+          isHidden: () => this.isHiddenIfNotRomsParser(),
           directory: true,
           label: this.lang.label.romDirectory,
           highlight: this.highlight.bind(this),
@@ -147,7 +148,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
           directory: true,
           label: this.lang.label.startInDirectory,
           highlight: this.highlight.bind(this),
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           onValidate: (self, path) => this.parsersService.validate(path[0] as keyof UserConfiguration, self.value),
           onInfoClick: (self, path) => {
             this.currentDoc.activePath = path.join();
@@ -156,14 +157,15 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
         }),
         userAccounts: new NestedFormElement.Group({
           label: this.lang.label.userAccounts,
-          isHidden: () => this.isHiddenIfNotAdvanced(),
           children: {
             specifiedAccounts: new NestedFormElement.Input({
               highlight: this.highlight.bind(this),
+              onValidateObservable: () => this.userForm.get('parserType').valueChanges,
               onValidate: (self, path) => {
                 let serialized = {}
                 serialized[path[1]] = self.value
-                return this.parsersService.validate(path[0] as keyof UserConfiguration, serialized)}
+                return this.parsersService.validate(path[0] as keyof UserConfiguration, serialized, {parserType: this.userForm.get('parserType').value});
+              }
             }),
             skipWithMissingDataDir: new NestedFormElement.Toggle({
               text: this.lang.text.skipWithMissingDataDir
@@ -216,7 +218,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
         })(),
 
         titleFromVariable: new NestedFormElement.Group({
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           label: this.lang.label.titleFromVariable,
           children: {
             limitToGroups: new NestedFormElement.Input({
@@ -253,7 +255,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
           }
         }),
         fuzzyMatch: new NestedFormElement.Group({
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           label: this.lang.label.fuzzyMatch,
           children: {
             use: new NestedFormElement.Toggle({
@@ -277,7 +279,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
         onlineImageQueries: new NestedFormElement.Input({
           label: this.lang.label.onlineImageQueries,
           highlight: this.highlight.bind(this),
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           onValidate: (self, path) => this.parsersService.validate(path[0] as keyof UserConfiguration, self.value),
           onInfoClick: (self, path) => {
             this.currentDoc.activePath = path.join();
@@ -287,7 +289,7 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
         imagePool: new NestedFormElement.Input({
           label: this.lang.label.imagePool,
           highlight: this.highlight.bind(this),
-          isHidden: () => this.isHiddenIfEither(),
+          isHidden: () => this.isHiddenIfAny(),
           onValidate: (self, path) => this.parsersService.validate(path[0] as keyof UserConfiguration, self.value),
           onInfoClick: (self, path) => {
             this.currentDoc.activePath = path.join();
@@ -477,15 +479,20 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
     this.currentDoc.activePath = '';
     this.currentDoc.content = this.lang.docs__md.communityPresets.join('');
   }
-
+  private isHiddenIfNotRomsParser() {
+    return Observable.concat(Observable.of(this.userForm.get('parserType').value),this.userForm.get('parserType').valueChanges).map(pType => !['Glob','Glob-regex'].includes(pType))
+  }
   private isHiddenIfNotAdvanced() {
     return Observable.concat(Observable.of(this.userForm.get('advanced').value), this.userForm.get('advanced').valueChanges).map(val => !val);
   }
   private isHiddenIfSteamParser() {
     return Observable.concat(Observable.of(this.userForm.get('parserType').value),this.userForm.get('parserType').valueChanges).map(pType => pType==='Steam');
   }
-  private isHiddenIfEither() {
-    return Observable.combineLatest(this.isHiddenIfNotAdvanced(),this.isHiddenIfSteamParser()).map(([val1,val2])=> val1||val2)
+  private isHiddenIfAny() {
+    return Observable.combineLatest(
+      this.isHiddenIfNotAdvanced(),
+      this.isHiddenIfSteamParser(),
+      this.isHiddenIfNotRomsParser()).map(([na,s,nr])=>na||s||nr)
   }
 
   private get lang() {
@@ -989,7 +996,6 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
     if (this.configurationIndex !== -1 && this.userConfigurations.length > this.configurationIndex) {
       let config = this.userConfigurations[this.configurationIndex];
       this.formChanges.unsubscribe();
-
       this.userForm.patchValue(config.current ? config.current : config.saved);
       this.markAsDirtyDeep(this.userForm);
 
@@ -1001,7 +1007,6 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
         else
           config.current = data;
       });
-
       this.loadedIndex = this.configurationIndex;
     }
     else if (this.configurationIndex === -1 && this.userConfigurations !== undefined) {
@@ -1019,7 +1024,6 @@ export class ParsersComponent implements AfterViewInit, OnDestroy {
 
   private markAsDirtyDeep(control: AbstractControl): void {
     control.markAsDirty();
-
     if (control['controls'] !== undefined) {
       for (let childKey in control['controls']) {
         this.markAsDirtyDeep(control['controls'][childKey]);