+++ src/renderer/modifiers/app-settings.modifier.ts
@@ -5,7 +5,7 @@ let versionUp = (version: number) => { return version + 1 };
 
 export const appSettings: ValidatorModifier<AppSettings> = {
   controlProperty: 'version',
-  latestVersion: 0,
+  latestVersion: 1,
   fields: {
     undefined: {
       'version': { method: () => 0 },
@@ -20,6 +20,17 @@ export const appSettings: ValidatorModifier<AppSettings> = {
             newValue[field] = newValue[field] || '';
           })
           return newValue;
+        }
+        return defaultValue;
+      }
+      }
+    },
+    0: {
+      'version': { method: versionUp },
+      'environmentVariables': { method: (oldValue) => {
+        let defaultValue = {retroarchPath:'',raCoresDirectory:'',steamDirectoryGlobal:'',localImagesDirectory:''}
+        if(oldValue){
+          return Object.assign(oldValue, {raCoresDirectory: ''});
         }
         return defaultValue;
       }